requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
pydantic
python-jose[cryptography]
bcrypt
cloudinary
redis
celery
python-decouple
httpx
python-dotenv
slowapi
sendgrid
twilio 